pyproject.toml
[project]
name = "portrait"
version = "1.0.0"
description = "用户数据画像服务 - 基于智能外呼系统数据"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Portrait Team" }
]

dependencies = [
    # Web 框架
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # 数据库
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0", # PostgreSQL 异步驱动
    "aiomysql>=0.2.0", # MySQL 异步驱动
    "alembic>=1.13.0", # 数据库迁移
    # 任务调度
    "apscheduler>=3.10.0",
    # 数据处理
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    # LLM
    "openai>=1.3.0",
    "httpx>=0.25.0",
    # 工具
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tenacity>=8.2.0", # 重试库
    "python-dateutil>=2.8.0",
    "cryptography>=46.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "pytest-asyncio>=1.3.0",
]

